SERVER-48411 Increase maximum allowed replica set memberId on disk from 255 to INT_MAX.
(cherry picked from commit 52db264df93fe1f5f75647136793bd025c0393a8)

+++ src/mongo/db/repl/repl_set_config.cpp
@@ -454,7 +454,7 @@ int ReplSetConfig::findMemberIndexByHostAndPort(const HostAndPort& hap) const {
     return -1;
 }
 
-int ReplSetConfig::findMemberIndexByConfigId(long long configId) const {
+int ReplSetConfig::findMemberIndexByConfigId(int configId) const {
     int x = 0;
     for (const auto& member : getMembers()) {
         if (member.getId() == MemberId(configId)) {
